+++ src/dialogs/settingsdialog.cpp
@@ -1168,10 +1168,16 @@ void SettingsDialog::on_noteFolderRemoveButton_clicked()
             0, 1) == 0) {
         bool wasCurrent = _selectedNoteFolder.isCurrent();
 
-        // remove saved searches
         QSettings settings;
+
+        // remove saved searches
         QString settingsKey = "savedSearches/noteFolder-"
                               + QString::number(_selectedNoteFolder.getId());
+        settings.remove(settingsKey);
+
+        // remove tree widget expand state setting
+        settingsKey = NoteSubFolder::treeWidgetExpandStateSettingsKey(
+                _selectedNoteFolder.getId());
         settings.remove(settingsKey);
 
         // remove the note folder from the database